% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignfigR.R
\docType{package}
\name{alignfigR}
\alias{alignfigR}
\alias{alignfigR-package}
\title{alignfigR.}
\description{
Visualizing multiple sequence alignments with ggplot.
CAUTION: no sanity checking is performed to check if your provided data is an alignment or just sequences (but they will plot either way :) )
}

