% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{basic_pal}
\alias{basic_pal}
\title{Defines basic Palette}
\format{
An object of class \code{character} of length 27.
}
\usage{
basic_pal
}
\description{
Defines basic Palette
}
\keyword{datasets}
