% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{calculate_column_percentage}
\alias{calculate_column_percentage}
\title{Reveals the unique protein/nucleotide identifiers found in this column, reveals the counts and the frequency of organisms having that identifier in that column}
\usage{
calculate_column_percentage(alignment_tibble, column_of_interest)
}
\arguments{
\item{alignment_tibble}{The tibble from read_alignment()}

\item{column_of_interest}{The specific column you want to see}
}
\value{
Returns a tibble with rows 'value', 'count' and 'percent'
}
\description{
Reveals the unique protein/nucleotide identifiers found in this column, reveals the counts and the frequency of organisms having that identifier in that column
}
