% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{calculate_total_nucs}
\alias{calculate_total_nucs}
\title{Calculates the total number of nucleotides in the data}
\usage{
calculate_total_nucs(data_longer)
}
\arguments{
\item{data_longer}{Tibble output from make_data_longer()}
}
\value{
Returns a sum of the number of nucleotides in the data
}
\description{
Calculates the total number of nucleotides in the data
}
