% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{calculate_total_seqs}
\alias{calculate_total_seqs}
\title{Calculates total seqs besides gaps in the data}
\usage{
calculate_total_seqs(data_longer)
}
\arguments{
\item{data_longer}{Tibble output from make_data_longer()}
}
\value{
Returns a sum of the number of proteins/nucleotides (ignoring gaps) in the data
}
\description{
Calculates total seqs besides gaps in the data
}
