% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{convert_seq_list_to_tibble}
\alias{convert_seq_list_to_tibble}
\title{Transforms seq_list into a tibble}
\usage{
convert_seq_list_to_tibble(seq_list)
}
\arguments{
\item{seq_list}{Dataframe output from read_alignment}
}
\value{
Returns a tibble of the fasta data
}
\description{
Transforms seq_list into a tibble
}
