% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_plot_functions.R
\name{create_alignment}
\alias{create_alignment}
\title{Creates the plot for plot_alignment}
\usage{
create_alignment(
  data_filtered,
  rect_alignment,
  pal,
  taxon_labels = FALSE,
  legend_title = "Legend Title",
  graph_title = "Graph Title"
)
}
\arguments{
\item{data_filtered}{The Output from filter_taxa_and_sites}

\item{rect_alignment}{The Output from create_rect_alignment}

\item{pal}{The output from define_palette}

\item{taxon_labels}{Determines if you would like rows to be labeled by their respective taxon}

\item{legend_title}{A String of what you want the legend title to be}

\item{graph_title}{A String of what you want the graph title to be}
}
\value{
Returns a Multiple Sequence Alignment
}
\description{
Creates the plot for plot_alignment
}
