% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_plot_functions.R
\name{create_geom_rect_alignment}
\alias{create_geom_rect_alignment}
\title{Creates a new tibble that is ready to be used in a geom_rect}
\usage{
create_geom_rect_alignment(data_longer)
}
\arguments{
\item{data_longer}{Output from make_data_longer}
}
\value{
Returns a tibble that contains columns x1, x2, y1, and y2; all with correct values for an alignment
}
\description{
Creates a new tibble that is ready to be used in a geom_rect
}
