% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_palette.R
\name{determine_palette}
\alias{determine_palette}
\title{Returns the palette specified by define_palette}
\usage{
determine_palette(data_longer, color_palette, custom_colors)
}
\arguments{
\item{data_longer}{Output from make_data_longer}

\item{color_palette}{The palette you wish to use. Options are "random", "dna", "rna", "custom", "free", "ocean", "fire", "forest" and "floral".}

\item{custom_colors}{A string of the colors you wish to have in the palette that contains the same amount of colors as unique protein/nucleotide identifiers you have in your data. The first identifier in uniques will be assigned the first color in custom_colors and so on.}
}
\value{
Returns a color palette
}
\description{
Returns the palette specified by define_palette
}
