% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{determine_type}
\alias{determine_type}
\title{Allows the package to determine data type, as well as allowing the user to specify this}
\usage{
determine_type(data, data_type = "")
}
\arguments{
\item{data}{Tibble output from convert_seq_list_to_tibble()}

\item{data_type}{The data_type associated with the data. If left blank, the function will determine the type of data}
}
\value{
Returns a tibble of the data that contains a column containing the type of data
}
\description{
Allows the package to determine data type, as well as allowing the user to specify this
}
