% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{determine_type_threshold}
\alias{determine_type_threshold}
\title{Defines the threshold for deteriming nucleotide of protein data type}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
determine_type_threshold
}
\description{
Defines the threshold for deteriming nucleotide of protein data type
}
\keyword{datasets}
