% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{extract_subalign}
\alias{extract_subalign}
\title{Gives the ability to select which taxa and columns you want portrayed in the plot}
\usage{
extract_subalign(
  alignment,
  tlist = c(),
  texcl = FALSE,
  clist = c(),
  cexcl = FALSE
)
}
\arguments{
\item{alignment}{The tibble from read_alignment()}

\item{tlist}{A string of the taxa you wish to include or exclude}

\item{texcl}{Determinant if you want to include only the taxa in tlist, or exclude only the tlist}

\item{clist}{The columns you wish to have portrayed in the plot}

\item{cexcl}{Determinant if you want to include only the columns in clist, or exclude only the clist}
}
\value{
Returns a tibble which contains the preferred taxa and columns, and is prepped for geom_rect()
}
\description{
Gives the ability to select which taxa and columns you want portrayed in the plot
}
