% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_and_determine_palette}
\alias{filter_and_determine_palette}
\title{Combines filter_and_make_data_longer with define_palette}
\usage{
filter_and_determine_palette(
  fasta_tibble,
  taxa,
  exclude_taxa,
  sites,
  exclude_sites,
  color_palette,
  custom_colors
)
}
\arguments{
\item{fasta_tibble}{Tibble output from read_alignment()}

\item{taxa}{List of desired or undesired taxa}

\item{exclude_taxa}{Determines if you wish to only include or exclude the taxa in 'taxa'}

\item{sites}{List of desired positions in the fasta sequence}

\item{exclude_sites}{Determines if you wish to only include or exclude the sites in 'sites'}

\item{color_palette}{The palette you wish to use. Options are "random", "dna", "rna", "custom", "free", "ocean", "fire", "forest" and "floral".}

\item{custom_colors}{A string of the colors you wish to have in the palette that contains the same amount of colors as unique protein/nucleotide identifiers you have in your data. The first identifier in uniques will be assigned the first color in custom_colors and so on.}
}
\value{
Returns a longer tibble of filtered_data
}
\description{
Combines filter_and_make_data_longer with define_palette
}
