% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_and_make_data_longer}
\alias{filter_and_make_data_longer}
\title{Combines filter_taxa_and_sites and make_data_longer}
\usage{
filter_and_make_data_longer(
  fasta_tibble,
  taxa,
  exclude_taxa,
  sites,
  exclude_sites
)
}
\arguments{
\item{fasta_tibble}{Tibble output from read_alignment()}

\item{taxa}{List of desired or undesired taxa}

\item{exclude_taxa}{Determines if you wish to only include or exclude the taxa in 'taxa'}

\item{sites}{List of desired positions in the fasta sequence}

\item{exclude_sites}{Determines if you wish to only include or exclude the sites in 'sites'}
}
\value{
Returns a longer tibble of filtered_data
}
\description{
Combines filter_taxa_and_sites and make_data_longer
}
