% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_taxa_and_sites}
\alias{filter_taxa_and_sites}
\title{Allows the user to select what taxa and sites are present in the graphs}
\usage{
filter_taxa_and_sites(
  fasta_tibble,
  taxa = c(),
  exclude_taxa = FALSE,
  sites = c(),
  exclude_sites = FALSE
)
}
\arguments{
\item{fasta_tibble}{Output from convert_seq_list_to_tibble}

\item{taxa}{List of desired or undesired taxa}

\item{exclude_taxa}{Determines if you wish to only include or exclude the taxa in 'taxa'}

\item{sites}{List of desired positions in the fasta sequence}

\item{exclude_sites}{Determines if you wish to only include or exclude the sites in 'sites'}
}
\value{
Returns a tibble of the fasta data with only the desired taxa and sites
}
\description{
Allows the user to select what taxa and sites are present in the graphs
}
