% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\docType{data}
\name{floral_pal}
\alias{floral_pal}
\title{Defines Floral Palette}
\format{
An object of class \code{character} of length 28.
}
\usage{
floral_pal
}
\description{
Defines Floral Palette
}
\keyword{datasets}
