% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\docType{data}
\name{forest_pal}
\alias{forest_pal}
\title{Defines Forest Palette}
\format{
An object of class \code{character} of length 27.
}
\usage{
forest_pal
}
\description{
Defines Forest Palette
}
\keyword{datasets}
