% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{make_data_longer}
\alias{make_data_longer}
\title{Pivots the output from filter_taxa_and_sites longer}
\usage{
make_data_longer(filtered_data)
}
\arguments{
\item{filtered_data}{The output from convert_seq_list_to_tibble}
}
\value{
Returns a longer tibble of filtered_data
}
\description{
Pivots the output from filter_taxa_and_sites longer
}
