% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{nucleotides}
\alias{nucleotides}
\title{Defines Nucleotides}
\format{
An object of class \code{character} of length 4.
}
\usage{
nucleotides
}
\description{
Defines Nucleotides
}
\keyword{datasets}
