% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{null_color}
\alias{null_color}
\title{Defines color for absence of protein/nucleotide}
\format{
An object of class \code{character} of length 1.
}
\usage{
null_color
}
\description{
Defines color for absence of protein/nucleotide
}
\keyword{datasets}
