% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ocean_pal}
\alias{ocean_pal}
\title{Defines Ocean Palette}
\format{
An object of class \code{character} of length 27.
}
\usage{
ocean_pal
}
\description{
Defines Ocean Palette
}
\keyword{datasets}
