% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Assigns the color palette for plot_alignment(), and allows you to select a custom palette}
\usage{
plot_alignment(
  alignment,
  tlist = c(),
  texcl = FALSE,
  clist = c(),
  cexcl = FALSE,
  typemsa,
  uniques = NA,
  custom_colors = NA,
  taxon_labels = FALSE,
  graph_title = "Graph Title",
  legend_title = "Legend"
)
}
\arguments{
\item{alignment}{The tibble from read_alignment()}

\item{tlist}{A string of the taxa you wish to include or exclude}

\item{texcl}{Determinant if you want to include only the taxa in tlist, or exclude only the tlist}

\item{clist}{The columns you wish to have portrayed in the plot}

\item{cexcl}{Determinant if you want to include only the columns in clist, or exclude only the clist}

\item{typemsa}{The palette you wish to use. Options are "random", "dna", "rna", "custom", "free", "ocean", "fire", "forest" and "floral".}

\item{uniques}{The protein/nucleotide identifiers used in your data}

\item{custom_colors}{A string of the colors you wish to have in the palette that contains the same amount of colors as unique protein/nucleotide identifiers you have in your data. The first identifier in uniques will be assigned the first color in custom_colors and so on.}

\item{taxon_labels}{Determinant of if taxa identifiers will be present on the graph or not}

\item{graph_title}{Sets the graph title}

\item{legend_title}{Sets the legend title}
}
\value{
Returns an MSA of the data
}
\description{
Assigns the color palette for plot_alignment(), and allows you to select a custom palette
}
