% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plot_functions.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{Creates the plot for plot_site_frequencies}
\usage{
plot_frequencies(
  output_from_prep_site_frequencies,
  pal,
  legend_title = "Legend Title",
  graph_title = "Graph Title"
)
}
\arguments{
\item{output_from_prep_site_frequencies}{The tibble output from prep_site_frequencies()}

\item{pal}{The output from define_palette}

\item{legend_title}{A String of what you want the legend title to be}

\item{graph_title}{A String of what you want the graph title to be}
}
\value{
Returns a plt of the site frequencies
}
\description{
Creates the plot for plot_site_frequencies
}
