% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_plot_functions.R,
%   R/frequency_plot_functions.R
\name{plot_site_frequencies}
\alias{plot_site_frequencies}
\title{Creates a Plot of Frequencies of protein identifiers at each site}
\usage{
plot_site_frequencies(
  fasta_tibble,
  taxa = c(),
  exclude_taxa = FALSE,
  sites = c(),
  exclude_sites = FALSE,
  color_palette,
  custom_colors = NA,
  legend_title = "Legend Title",
  graph_title = "Graph Title"
)

plot_site_frequencies(
  fasta_tibble,
  taxa = c(),
  exclude_taxa = FALSE,
  sites = c(),
  exclude_sites = FALSE,
  color_palette,
  custom_colors = NA,
  legend_title = "Legend Title",
  graph_title = "Graph Title"
)
}
\arguments{
\item{fasta_tibble}{Output from convert_seq_list_to_tibble}

\item{taxa}{List of desired or undesired taxa}

\item{exclude_taxa}{Determines if you wish to only include or exclude the taxa in 'taxa'}

\item{sites}{List of desired positions in the fasta sequence}

\item{exclude_sites}{Determines if you wish to only include or exclude the sites in 'sites'}

\item{color_palette}{The palette you wish to use. Options are "random", "dna", "rna", "custom", "free", "ocean", "fire", "forest" and "floral".}

\item{custom_colors}{A string of the colors you wish to have in the palette that contains the same amount of colors as unique protein/nucleotide identifiers you have in your data. The first identifier in uniques will be assigned the first color in custom_colors and so on.}

\item{legend_title}{A String of what you want the legend title to be}

\item{graph_title}{A String of what you want the graph title to be}
}
\value{
Returns a Plot of Frequencies of protein identifiers at each site

Returns a Plot of Frequencies of protein identifiers at each site
}
\description{
Creates a Plot of Frequencies of protein identifiers at each site

Creates a Plot of Frequencies of protein identifiers at each site
}
