% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plot_functions.R
\name{prep_site_frequencies}
\alias{prep_site_frequencies}
\title{Allows the user to select what taxa and sites are present in the graphs}
\usage{
prep_site_frequencies(data_longer)
}
\arguments{
\item{data_longer}{Tibble output from make_data_longer()}
}
\value{
Returns a tibble that is ready for plot_frequencies()
}
\description{
Allows the user to select what taxa and sites are present in the graphs
}
