% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\docType{data}
\name{protein_default_pal}
\alias{protein_default_pal}
\title{Defines basic Palette}
\format{
An object of class \code{character} of length 27.
}
\usage{
protein_default_pal
}
\description{
Defines basic Palette
}
\keyword{datasets}
