% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\docType{data}
\name{protein_fire_pal}
\alias{protein_fire_pal}
\title{Defines Fire Palette}
\format{
An object of class \code{character} of length 28.
}
\usage{
protein_fire_pal
}
\description{
Defines Fire Palette
}
\keyword{datasets}
