% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\docType{data}
\name{protein_ocean_pal}
\alias{protein_ocean_pal}
\title{Defines Ocean Palette}
\format{
An object of class \code{character} of length 27.
}
\usage{
protein_ocean_pal
}
\description{
Defines Ocean Palette
}
\keyword{datasets}
