% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_alignment}
\alias{read_alignment}
\title{Transforms fasta into a tibble}
\usage{
read_alignment(file, data_type = "")
}
\arguments{
\item{file}{The fasta}

\item{data_type}{The data_type associated with the data. If left blank, the function will determine the type of data}
}
\value{
Returns a tibble of the fasta data
}
\description{
Transforms fasta into a tibble
}
