% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_plot_functions.R,
%   R/frequency_plot_functions.R
\name{create_site_frequencies}
\alias{create_site_frequencies}
\title{Creates the plot for plot_site_frequencies}
\usage{
create_site_frequencies(
  data_longer,
  pal,
  legend_title = "Legend Title",
  graph_title = "Legend Title"
)

create_site_frequencies(
  data_longer,
  pal,
  legend_title = "Legend Title",
  graph_title = "Legend Title"
)
}
\arguments{
\item{data_longer}{The Output from make_data_longer}

\item{pal}{The output from define_palette}

\item{legend_title}{A String of what you want the legend title to be}

\item{graph_title}{A String of what you want the graph title to be}
}
\value{
Returns a Plot of Frequencies of protein identifiers at each site

Returns a Plot of Frequencies of protein identifiers at each site
}
\description{
Creates the plot for plot_site_frequencies

Creates the plot for plot_site_frequencies
}
