% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{define_palette}
\alias{define_palette}
\title{Assigns the color palette for plot_alignment(), and allows you to select a custom palette}
\usage{
define_palette(palette_msa, uniques = NA, custom_colors = NA)
}
\arguments{
\item{palette_msa}{The palette you wish to use. Options are "random", "dna", "rna", "custom", "free", "ocean", "fire", "forest" and "floral".}

\item{uniques}{The protein/nucleotide identifiers used in your data}

\item{custom_colors}{A string of the colors you wish to have in the palette that contains the same amount of colors as unique protein/nucleotide identifiers you have in your data. The first identifier in uniques will be assigned the first color in custom_colors and so on.}
}
\value{
Returns a color palette
}
\description{
Assigns the color palette for plot_alignment(), and allows you to select a custom palette
}
