% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{extract_subalign}
\alias{extract_subalign}
\title{Extract subset of sequence alignment}
\usage{
extract_subalign(seqs, s, tlist, clist, taction, caction)
}
\arguments{
\item{seqs}{Sequence array}

\item{s}{Step size for alignment block. Default, 20.}

\item{tlist}{List of taxa (the actual labels, not order) to either restrict to or exclude from plot.}

\item{clist}{List of columns (indexed from 1) to either restrict to or exclude from plot.}

\item{taction}{"exclude" or "restrict". If "exclude", taxa given in taxon_list will not be plotted. If "restrict", *only* the taxa given in taxon_list will be plotted. DEFAULT: exclude.}

\item{caction}{"exclude" or "restrict". If "exclude", column given in column_list will not be plotted. If "restrict", *only* the column given in column_list will be plotted. DEFAULT: exclude.}
}
\value{
plot_frame, a data frame to be plotted
}
\description{
This function builds a data frame to plot an alignment from a specified subset of the full alignment.
}
\examples{
GIVE EXAMPLES
}

