% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{extract_subalign}
\alias{extract_subalign}
\title{Extract subset of sequence alignment}
\usage{
extract_subalign(seqs, s, tlist, clist, texcl, cexcl)
}
\arguments{
\item{seqs}{Sequence array}

\item{s}{Step size for alignment block. Default, 1.}

\item{tlist}{List of taxa (the actual labels, not order) to either restrict to or exclude from plot.}

\item{clist}{List of columns (indexed from 1) to either restrict to or exclude from plot.}

\item{texl}{Exclude those taxa?}

\item{cexl}{Exclude those columns?}
}
\value{
plot_frame, a data frame to be plotted
}
\description{
This function builds a data frame to plot an alignment from a specified subset of the full alignment.
}
\examples{
GIVE EXAMPLES
}

