% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{extract_subalign}
\alias{extract_subalign}
\title{Extract subset of sequence alignment}
\usage{
extract_subalign(seqs, s, tlist, clist, texcl, cexcl)
}
\arguments{
\item{seqs}{Sequence array}

\item{s}{Step size for alignment block. Default, 1.}

\item{tlist}{List of taxa intended to restrict figure to or to exclude from figure.}

\item{clist}{List of columns intended to restrict figure to or to exclude from figure.}

\item{texl}{Boolean indicating if taxa in tlist should be exluded.}

\item{cexl}{Boolean indicating if columns in clist should be exluded.}
}
\value{
plot_frame, a data frame to be plotted
}
\description{
This function builds a data frame to plot an alignment from a specified subset of the full alignment.
}
\examples{
GIVE EXAMPLES
}

