% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{nucleotide_pal}
\alias{nucleotide_pal}
\title{Needs a thing}
\format{
An object of class \code{character} of length 5.
}
\usage{
nucleotide_pal
}
\description{
Needs a thing
}
\keyword{datasets}
