% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignfigR.R
\name{read_alignment}
\alias{read_alignment}
\title{Read a multiple sequence alignment file.}
\usage{
read_alignment(file)
}
\arguments{
\item{file}{File name. NOTE: Only FASTA file are supported!}
}
\value{
seq_array, a named-array of the parsed sequence data
}
\description{
This function parses a FASTA file containing molecular sequence data and returns a named-array of the sequences.
}
\examples{
fasta_file <- system.file("extdata", "example.fasta", package = "alignfigR")
plot_frame <- read_alignment(file = fasta_file)
}
