% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_alignment}
\alias{read_alignment}
\title{Transforms fasta into a tibble}
\usage{
read_alignment(file)
}
\arguments{
\item{file}{The fasta}
}
\value{
Returns a tibble of the fasta data
}
\description{
Transforms fasta into a tibble
}
